requirements.txt
discord.py~=2.3.2
pymongo~=4.6.2 
